Minor refactoring and update changelog
